--- a/aeon/clustering/tests/test_k_medoids.py
+++ b/aeon/clustering/tests/test_k_medoids.py
@@ -133,3 +133,20 @@ def test_kmedoids():
 
     for val in proba:
         assert np.count_nonzero(val == 1.0) == 1
+
+def test_new_kmedoids():
+    X_train, y_train = load_basic_motions(split="train")
+    X_test, y_test = load_basic_motions(split="test")
+
+    kmedoids = TimeSeriesKMedoids(
+        random_state=1,
+        n_init=2,
+        max_iter=5,
+        init_algorithm="kmeans++",
+        distance="euclidean",
+    )
+    train_predict = kmedoids.fit_predict(X_train)
+    train_score = metrics.rand_score(y_train, train_predict)
+    test_medoids_result = kmedoids.predict(X_test)
+    medoids_score = metrics.rand_score(y_test, test_medoids_result)
+    proba = kmedoids.predict_proba(X_test)