--- a/aeon/clustering/tests/test_k_medoids.py
+++ b/aeon/clustering/tests/test_k_medoids.py
@@ -114,8 +114,8 @@ def test_kmedoids():
         random_state=1,
         n_init=2,
         max_iter=5,
-        init_algorithm="kmeans++",
-        metric="euclidean",
+        init_algorithm="random",
+        distance="euclidean",
     )
     train_predict = kmedoids.fit_predict(X_train)
     train_score = metrics.rand_score(y_train, train_predict)