<!--
Thanks for contributing a pull request! Please ensure you have taken a look at our contribution guide: https://github.com/aeon-toolkit/aeon/blob/main/CONTRIBUTING.md

Feel free to delete sections of this template if they do not apply to your PR, avoid submitting a blank template or empty sections.
-->

#### Reference Issues/PRs

<!--
Example: Fixes #1234. See also #3456.

Please use keywords (e.g., Fixes) to create link to the issues or pull requests you resolved, so that they will automatically be closed when your pull request is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->

#### What does this implement/fix? Explain your changes.
Kmedoids was attached to the lloyds partioning algorithm which was implemented and then inherited and used for both kmedoids and kmeans. One of the issue with this is it obfuscated the code for kmedoids but also made it very hard to makes medoids specific changes to the algorithm. This PR splits up kmedoids from the lloyds abstact class and reimplements the logic so it is stand alone.

This means medoids specific optimisations and inits can now be added without causing issues with kmeans. Additionally splitting allows for new medoids variants to more easily be added in the future (e.g. PAM).

<!--
A clear and concise description of what you have implemented.
-->

#### Does your contribution introduce a new dependency? If yes, which one?
No

<!--
If your contribution does add a dependency, we may suggest adding it as an optional/soft dependency to keep external dependencies of the core aeon package to a minimum.
-->

#### What should a reviewer concentrate their feedback on?

<!--
This section is particularly useful if you have a pull request that is still in development. You can guide the reviews to focus on the parts that are ready for their comments. We suggest using bullets (indicated by * or -) and filled checkboxes [x] here
-->

#### Did you add any tests for the change?
Yes

<!--
This section is useful if you have added a test in addition to the existing ones. This will ensure that further changes to these files won't introduce the same kind of bug. It is considered good practice to add tests with newly added code to enforce the fact that the code actually works. This will reduce the chance of introducing logical bugs.
-->

#### Any other comments?
This PR ensures equivalence to the previous kmedoids implementation. Additionally it adds more comprhensive testing for the algorithm and init.

In addition significant speeds up have been added see the below table:

The kmedoids timing experiment was run using default parameters and DTW distance

| Dataset  | Previous kmedoids implementation speed (seconds) | New kmedoids implementation speed (seconds)|
| ------------- | ------------- | ------------- |
| Gunpoint | 13.293811  | 1.142850  |
| BasicMotions | 3.611624  | 0.253455  |
| ItalyPowerDemand | 0.60075  | 0.200906  |

This has been achieved by using a caching version of medoids suggested in the fasterPAM paper.

<!--
Please be aware that we are a loose team of volunteers so patience is necessary; assistance handling other issues is very welcome. We value all user contributions, no matter how minor they are. If we are slow to review, either the pull request needs some benchmarking, tinkering, convincing, etc. or more likely the reviewers are simply busy. In either case, we ask for your understanding during the review process.
-->

### PR checklist

<!--
Please go through the checklist below. Please feel free to remove points if they are not applicable.
-->

##### For all contributions
- [x] I've added myself to the [list of contributors](https://github.com/aeon-toolkit/aeon/blob/main/.all-contributorsrc).
- [x] Optionally, I've updated aeon's [CODEOWNERS](https://github.com/aeon-toolkit/aeon/blob/main/CODEOWNERS) to receive notifications about future changes to these files.
- [x] The PR title starts with either [ENH], [MNT], [DOC], or [BUG] indicating whether the PR topic is related to enhancement, maintenance, documentation, or bug.
- [x] For new estimators, I've added the estimator to the online documentation.


<!--
Thanks for contributing!
-->