@chrisholder 2023-07-05
I have added examples in the following PR for PAM (easier to do together)